requirements.txt
fastapi
uvicorn
httpx
pytest
motor
pytest-cov
pytest-asyncio
python-jose[cryptography]
python-multipart